requirements.txt
# Core ML dependencies
torch>=1.12.0,<2.0.0
torchvision>=0.13.0,<1.0.0
onnx>=1.12.0
onnxruntime>=1.12.0

# Image processing
Pillow>=9.0.0
opencv-python-headless>=4.6.0
numpy>=1.21.0

# Web and API
requests>=2.28.0

# Utilities
psutil>=5.9.0
pathlib2>=2.3.7

# Development and testing (optional, but good for debugging)
pytest>=7.0.0